requirements.txt
# requirements.txt
espnet
pyopenjtalk-prebuilt==0.2.0
typeguard==2.13.3
espnet_model_zoo
torchaudio
ipykernel
fastapi
jinja2
uvicorn[standard]
pydub
beautifulsoup4
-e ../py_code/whisper
sudachipy
sudachidict_core
neologdn
langchain
ginza
ja_ginza
pyautogui